\name{coef-methods}
\docType{methods}
\alias{coef-methods}
\alias{coef,ANY-method}
\title{ Methods for Function coef }
\description{ Methods for function \code{coef}}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{ default coef function ?}

\item{\code{signature(x = "SaemixObject")}}{ extracts coefficients from an SaemixObject}

}}
\keyword{methods}
