\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\title{ Methods for Function initialize}
\description{ Constructor functions for Classes in the saemix package (not user-level) }
\section{Methods}{
\describe{

\item{\code{signature(.Object = "SaemixData")}}{ create a SaemixData object. Please use the \code{\link{saemixData}} function.}

\item{\code{signature(.Object = "SaemixModel")}}{ create a SaemixModel object Please use the \code{\link{saemixModel}} function.}

\item{\code{signature(.Object = "SaemixObject")}}{ create a SaemixObject object. This object is obtained after a successful call to \code{\link{saemix}}}

\item{\code{signature(.Object = "SaemixRepData")}}{ create a SaemixRepData object}

\item{\code{signature(.Object = "SaemixRes")}}{ create a SaemixRes object}

\item{\code{signature(.Object = "SaemixSimData")}}{ create a SaemixSimData object}
}}
\keyword{methods}
