\name{show-methods}
\docType{methods}
\alias{show-methods}
\title{ Methods for Function show }
\description{ Prints a short summary of an object}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
Default show function
}

\item{\code{signature(x = "SaemixData")}}{
Prints a short summary of a SaemixData object
}

\item{\code{signature(x = "SaemixModel")}}{
Prints a short summary of a SaemixModel object
}

\item{\code{signature(x = "SaemixObject")}}{
Prints a short summary of the results from a SAEMIX fit
}

\item{\code{signature(x = "SaemixRes")}}{
Not user-level
}

\item{\code{signature(object = "SaemixRepData")}}{
Prints a short summary of a SaemixRepData object
}

\item{\code{signature(object = "SaemixSimData")}}{
Prints a short summary of a SaemixSimData object
}
}}
\keyword{methods}
