\name{showall-methods}
\docType{methods}
\alias{showall-methods}
\title{ Methods for Function showall}
\description{ Prints an extensive summary of an object}
\section{Methods}{
\describe{
\item{\code{signature(x = "SaemixData")}}{
Prints a extensive summary of a SaemixData object
}

\item{\code{signature(x = "SaemixModel")}}{
Prints a extensive summary of a SaemixModel object
}

\item{\code{signature(x = "SaemixObject")}}{
Prints a extensive summary of the results from a SAEMIX fit
}

\item{\code{signature(x = "SaemixRes")}}{
Not user-level
}
}}
\keyword{methods}
