\name{[-methods}
\docType{methods}
\alias{[-methods}
\title{ Methods for "get" Function [ }
\description{ Methods for function \code{[}, to allow access to the value of an element within an S4 object}
\section{Methods}{
\describe{

\item{\code{signature(x = "SaemixData")}}{ access elements of an object SaemixData}

\item{\code{signature(x = "SaemixModel")}}{ access elements of an object SaemixModel}

\item{\code{signature(x = "SaemixObject")}}{ access elements of an object SaemixObject}

\item{\code{signature(x = "SaemixRepData")}}{ access elements of an object SaemixRepData}

\item{\code{signature(x = "SaemixRes")}}{ access elements of an object SaemixRes}

\item{\code{signature(x = "SaemixSimData")}}{ access elements of an object SaemixSimData}
}}
\keyword{methods}
