\name{[<--methods}
\docType{methods}
\alias{[<--methods}
\title{Methods for "set" Function [<-}
\description{Methods for function \code{[<-}, to allow replacements of the value of an element within an S4 object}
\section{Methods}{
\describe{

\item{\code{signature(x = "SaemixData")}}{ replace elements of an object SaemixData}

\item{\code{signature(x = "SaemixModel")}}{ replace elements of an object SaemixModel}

\item{\code{signature(x = "SaemixObject")}}{ replace elements of an object SaemixObject}

\item{\code{signature(x = "SaemixRepData")}}{ replace elements of an object SaemixRepData}

\item{\code{signature(x = "SaemixRes")}}{ replace elements of an object SaemixRes}

\item{\code{signature(x = "SaemixSimData")}}{ replace elements of an object SaemixSimData}
}}
\keyword{methods}
